[CUBRIDQA-71] fix CUBRID environment variables (#258)
* [CUBRIDQA-71] fix CUBRID environment variables

* [CUBRIDQA-71] fix the path problem of the excluded file

* [CUBRIDQA-71] fix some rm statements for security operation

* [CUBRIDQA-71] print more information for deployment

+++ CTP/common/ext/run_jdbc.sh
@@ -79,7 +79,9 @@ function run_jdbc {
     cp -rf current_runtime_logs $backup_file
     cp $ctp_jdbc_test_conf $backup_file
     tar zvcf ${backup_file}.tar.gz $backup_file
-    rm -rf $backup_file
+    if [ "${backup_file}" ];then
+    	rm -rf $backup_file
+    fi
 }
